pyproject.toml
[project]
name = "agenticrag"
version = "0.1.0"
description = ""
authors = [
    {name = "AmandaZ",email = "njuzhangyujie@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langgraph (>=0.5.4,<0.6.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "tavily-python (>=0.7.10,<0.8.0)",
    "chromadb (>=1.0.15,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "langchain-chroma (>=0.2.5,<0.3.0)",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "langchain-tavily (>=0.2.11,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."

